# Weighted blend memory reduction