--- a/improver/blending/blend_across_adjacent_points.py
+++ b/improver/blending/blend_across_adjacent_points.py
@@ -32,7 +32,6 @@
 opposed to collapsing the whole dimension."""
 
 import iris
-import numpy as np
 from cf_units import Unit
 from iris.cube import CubeList