--- a/improver/blending/calculate_weights_and_blend.py
+++ b/improver/blending/calculate_weights_and_blend.py
@@ -242,7 +242,6 @@ class WeightAndBlend(BasePlugin):
 
             # blend across specified dimension
             BlendingPlugin = WeightedBlendAcrossWholeDimension(self.blend_coord)
-
             result = BlendingPlugin(
                 cube,
                 weights=weights,