--- a/improver/blending/weighted_blend.py
+++ b/improver/blending/weighted_blend.py
@@ -720,7 +720,11 @@ class WeightedBlendAcrossWholeDimension(PostProcessingPlugin):
 
         (collapse_dim,) = cube.coord_dims(self.blend_coord)
         slice_dim = 1 if collapse_dim == 0 else 0
-        allow_slicing = cube.ndim > 1 and cube.shape[slice_dim] > 1
+
+        slicing_threshold = 1620000  # (2, 900, 900)
+        allow_slicing = (
+            cube.ndim > 2 and np.array(cube.shape).prod() > slicing_threshold
+        )
 
         cube_slices = cube.slices_over(slice_dim) if allow_slicing else [cube]
         weights_slices = (
@@ -733,9 +737,6 @@ class WeightedBlendAcrossWholeDimension(PostProcessingPlugin):
             collapsed(c_slice, self.blend_coord, iris.analysis.MEAN, weights=w_slice)
             for c_slice, w_slice in zip(cube_slices, weights_slices)
         )
-        self.has_mask = False
-        for c_slice in cube_slices:
-            if isinstance(c_slice.data, np.ma.core.MaskedArray): self.has_mask = True
 
         result = result_slices.merge_cube() if allow_slicing else result_slices[0]
 
@@ -913,9 +914,5 @@ class WeightedBlendAcrossWholeDimension(PostProcessingPlugin):
         else:
             result = self.weighted_mean(cube, weights)
         self._update_blended_metadata(result, attributes_dict)
-
-        # Re-mask output
-        if self.has_mask:
-            result.data = np.ma.array(result.data)
 
         return result