Addresses #1246 

Reducing the memory requirements for weighted blending by slicing the data and not creating large weights arrays.

## Benchmarking 

Feels like temperature:
 - Before:
   - 5148.673 MB
   - 26.708 s
 - After
   - 225.746 MB
   - 33.633 s

Blend adjacent points:
 - Before:
   - 1006.643 MB
   - 11.041 s
 - After
   - 312.639 MB
   - 10.754 s

Testing:
 - [x] Ran tests and they passed OK
 - [x] Added new tests for the new feature(s)