@benfitzpatrick 2020-05-21
![Codacy](https://app.codacy.com/assets/images/favicon.png) Here is an overview of what got changed by this pull request:


```diff

Complexity increasing per file
==============================
- improver/blending/blend_across_adjacent_points.py  2
- improver/blending/calculate_weights_and_blend.py  8
         
```


See the complete overview on [Codacy](https://app.codacy.com/manual/metoppv_tech/improver/pullRequest?prid=5556738&bid=18223349)