@cgsandford 2020-05-22
@TomekTrzeciak I've put in a comment on the slicing conditions which I think should catch the cases that might trip us up.  The underpinning assumption is that cube dimensions will always be ordered with the spatial dimensions last, which we enforce on cube loading, so I believe that is safe.  I can't think of any other problems immediately.