@TomekTrzeciak 2020-05-22
> I don't think you would ever want to slice over a spatial dimension

We do in some cases - #1212 is a good example (slicing arrays rather than cubes)

> (although I suppose technically you could, I can't see that it would break anything in this instance). I can't think of any other problems immediately.

This would actually break the threshold code due to auto-demotion business of length 1 dims to scalar coords on merge in Iris (I tried).